fix(web): 🐛 bountyContract store useRequest update
